invoice: Add list items route
